requirements.txt
pymongo==4.8.0
python-dotenv==1.0.1
flask==2.2.3
flask-cors==4.0.0
flask_login==0.6.2
itsdangerous==2.2.0
gunicorn==20.1.0
Werkzeug==2.2.3
flask_mail
requests==2.25.1
ua-parser==1.0.1
special-agents
user_agents
geoip2
re==2.2.1